% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_summed_exps.R
\name{log_summed_exps}
\alias{log_summed_exps}
\title{Logged Summation of Exponentiated Vector Elements}
\usage{
log_summed_exps(x)
}
\arguments{
\item{x}{A vector of type numeric}
}
\value{
The logged summation of exponentiated input vector
}
\description{
Logged Summation of Exponentiated Vector Elements
}
\examples{
log_summed_exps(1:2000)
log_summed_exps(seq(1, 4000, by = 2))
log_summed_exps(c(2000, 50, 30, 1, 5, 7))

}
